+++ lib/rom/sql/relation/reading.rb
@@ -1011,20 +1011,6 @@ module ROM
           new(dataset.__send__(__method__))
         end
 
-        # Returns a relation that will never issue a query to the database. It
-        # implements the null object pattern for relations.
-        #
-        # @see http://sequel.jeremyevans.net/rdoc-plugins/files/lib/sequel/extensions/null_dataset_rb.html
-        # @example result will always be empty, regardless if records exists
-        #   users.where(name: 'Alice').nullify
-        #
-        # @return [SQL::Relation]
-        #
-        # @api public
-        def nullify
-          new(dataset.__send__(__method__))
-        end
-
         private
 
         # Build a locking clause